% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_data.R
\docType{data}
\name{ad_data}
\alias{ad_data}
\title{Ad Data}
\format{
A data frame with 88 rows and 3 variables:
\describe{
\item{rating}{Rating score from the four -3 to 3 scales. (Quantitative Discrete -12,through 12)}
\item{claim}{Type of claim ("concrete" or "abstract")}
\item{imagery}{Type of imagery delivered ("Verbal" or "Visual")}
}
}
\source{
Rossiter, J. R., & Percy, L. (1978). Visual imaging ability as a mediator of advertising response. Advances in Consumer Research, 5(1), 621-629.
}
\usage{
ad_data
}
\description{
In this study, the researchers wanted to focus on the type of visual imagery and how that might mitigate any effects of the type of claim (concrete vs. abstract). The ad for one brand was presented with a large picture of the product and a small verbal copy (words) underneath (“strong visual”). This strong visual ad was then paired with both the concrete and abstract claims. The ad for the other brand was presented with a large verbal copy and a small picture underneath (“strong verbal”). This strong verbal ad was paired with both the concrete and abstract claims. Note: The researchers chose not to use a control ad with no picture because they felt a beer ad with no pictures would not be very realistic. After viewing the ad, consumers were asked to rate their attitude toward the brand using four −3 to 3 scales (so total scores can range from −12 to 12). Participants were 44 men and 44 women recruited from a typical Midwestern suburban shopping center. All respondents were beer drinkers, and attempts were made to recruit individuals of a variety of ages.
}
\keyword{datasets}
