% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biodiesel.R
\docType{data}
\name{Biodiesel}
\alias{Biodiesel}
\title{Biodiesel Dataset}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{yield}{Yield percentage of methyl ester produced. (Quantitative, percentage.)}
\item{Temperature}{Temperature at which the reaction was preformed. (Quantitative, in Celcius.)}
\item{Concentration}{Amount of catalyst present at start of reaction, as a concentration. (Quantitative, weight percentage.)}
}
}
\source{
Vicente, G. et al., 1998, “Application of the factorial design of experiments and response sur-face methodology to optimize biodiesel production,” International Journal of Industrial Crops and Products, pp. 29−35.
}
\usage{
Biodiesel
}
\description{
Study investigating the process of using a catalyst (a chemical) to convert the triglycerides in sunflower oil to methyl ester (ME), the fuel source in biodiesel. Two factors were studied: the temperature at which the chemical reaction was performed (°C) and the amount of catalyst present (as a concentration, weight\%).
}
\keyword{datasets}
