% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CherData.R
\docType{data}
\name{CherData}
\alias{CherData}
\title{CherData Dataset}
\format{
A data frame with 4725 rows and 3 variables:
\describe{
\item{live.chernobyl}{Binary variable for whether or not respondent ever lived in the Chernobyl contaminated zone. ("No" or "Yes")}
\item{age}{Age of participant. (Quantitative, discrete)}
\item{sex}{Sex of participant, ("male" or "female")}
\item{alc.post}{Binary variable for whether or not respondent developed alcohol abuse or dependence after the Chernobyl disaster. ("yes" or "no" or "before")}
}
}
\source{
Bolt, M.A., Helming, L.M., & Tintle, N.L. (2018). The associations between self-reported exposure to the Chernobyl nuclear disaster zone and mental health disorders in Ukraine. Frontiers in Psychiatry, 9, 32.
}
\usage{
CherData
}
\description{
In a recent study, researchers investigated rates of alcohol abuse among men and women in Ukraine, investigating questions about whether rates differ depending on variables such as sex and age, and after adjusting for such potential confounding variables, how the rates may be related to exposure to different conflicts and traumas, in particular living near the Chernobyl nuclear power plant at the time of the accident.
}
\keyword{datasets}
