% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEV.R
\docType{data}
\name{FEV}
\alias{FEV}
\title{FEV Dataset}
\format{
A data frame with 654 rows and 5 variables:
\describe{
\item{Age}{Age of participant (Quantitative discrete)}
\item{FEV}{Forced expiratory volume, measure of strength of lungs. (Quantitative, max volume of air in liters a person can blow in first second.)}
\item{Height}{Height of participant (Quantitative, inches, rounded to half inch. )}
\item{Gender}{Gender of participant, ("Male" or "Female" )}
\item{Smoker}{If participant smokes. ("yes" or "no")}
}
}
\source{
Tager, Weiss, Rosner, & Speizer, 1979, “Effect of parental cigarette smoking on the pulmonary function of children,” American Journal of Epidemiology
}
\usage{
FEV
}
\description{
Forced expiratory volume (FEV) is a measure of the strength of a person’s lungs—the maximum volume of air (measured in liters) a person can blow out in the first second. Larger values indicate healthier lungs. Researchers in the 1970s were interested in how smoking impacted FEV in adolescents.
Researchers were specifically focused on if smoking has a particularly strong impact on adolescents whose lung capacity is still developing and maturing.
}
\keyword{datasets}
