% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\docType{data}
\name{height}
\alias{height}
\title{height Dataset}
\format{
A data frame with 174 rows and 3 variables:
\describe{
\item{hopelessness8}{Scale of hopelessness 8 moths after discharged from hospital. (Quantitative 1-4)}
\item{Exercise}{Level of exercise ("exerciser" or "non-exerciser")}
\item{BaselineGroup}{Level of hopelessness when initially released, ("mod_severeHopeless" or "notHopeless" )}
}
}
\source{
Dunn, S. L., Dunn, L. M., Rieth, N. P., Olamijulo, G. B., Swieringa, L. L., Holden, T. P., Clark, J. A., DeVon, H. A., & Tintle, N. L. (2017). Impact of home- and hospital-based exercise in cardiac rehabilitation on hopelessness in patients with coronary heart disease. Journal of Cardiopulmonary Rehabilitation and Prevention, 37(1), 39-48. https://doi.org/10.1097/HCR.0000000000000205
}
\usage{
height
}
\description{
Study on participants with recent heart attack requiring hospitalization. Participants completed surveys before being discharged from the hospital. Chronic hopelessness was measured on a series of Linkert Scales, which were averaged into a Hopelessness scale 1-4. Researchers were interested in the effect of exercise on hopelessness levels eight months later.
}
\keyword{datasets}
