% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_data.R
\docType{data}
\name{housing_data}
\alias{housing_data}
\title{housing_data Dataset}
\format{
A data frame with 13 rows and 3 variables:
\describe{
\item{sqft}{Square footage of house. (Quantitative, rounded to square foot.)}
\item{price.1000}{Price of house in thousands of dollars. (Quantitative, rounded to hundred dollars. )}
\item{lake}{Binary categorical variable of if the house is on a lake or not. ("lakefront" or "notlakefront")}
}
}
\source{
Tintle, N., Chance, B., Cobb, G., Rossman, A., Roy, S., & VanderStoep, J. (2019). Intermediate Statistical Investigations. John Wiley & Sons. housing prices in Michigan Dataset.
}
\usage{
housing_data
}
\description{
In 2015, we sampled 13 homes that were for sale just north of Lake Macatawa in Michigan, and recorded the sale price (in thousands of dollars) as listed on Zillow (a real estate website).
}
\keyword{datasets}
