% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricanesA.R
\docType{data}
\name{hurricanesA}
\alias{hurricanesA}
\title{Hurricane Dataset}
\format{
A data frame with 20 rows and 3 variables:
\describe{
\item{Rating}{Rating of riskiness (7 point Linkert Scale)}
\item{HurrGender}{Gender of name ascribed to hurricane ("male" or "female")}
\item{Participant}{Not included in original study. Participant ID ("1" or "346")}
}
}
\source{
Jung, K., Shavitt, S., Viswanathan, M., & Hilbe, J. M. (2014). Female hurricanes are deadlier than male hurricanes. Proceedings of the National Academy of Sciences of the United States of America, 111(24), 8782–8787. https://doi.org/10.1073/pnas.1402786111
}
\usage{
hurricanesA
}
\description{
A study into perceived level of danger and hurricane name. Participants were given a weather map showing a hurricane moving toward the area where they lived. Participants were told the name of the hurricane and then rated the riskiness of the hurricane on a seven-point scale, with 1 being “not at all” (risky) and 7 being “very risky." This dataset only contains two participants, ID = 1 and  ID = 346.
}
\keyword{datasets}
