% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infant.R
\docType{data}
\name{Infant}
\alias{Infant}
\title{Infant Dataset}
\format{
A data frame with 49 rows and 2 variables:
\describe{
\item{Attatchment}{Level of attachment after 12 months. ("Secure Attachment" or "Insecure Attachment")}
\item{Device}{Device or baby carrier assigned to the mother. ("Soft Baby Carriers" or "Plastic Infant Seats")}
}
}
\source{
Anisfeld, Elizabeth, L. Casper, I. Nozyce, & M. Cunningham. (1990). “Does Infant Carrying Promote Attachment?”  Child Development, 61(5), 1617–1627. https://doi.org/10.2307/1130767
}
\usage{
Infant
}
\description{
A psychological study where low socioeconomic status mothers were recruited to use one of two different types of infant carriers. Soft baby carriers or Hard plastic baby carriers were given and the attachment outcome was recorded 12 months later using the Ainsworth Strange Situation procedure.
}
\keyword{datasets}
