% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYDerby18.R
\docType{data}
\name{KYDerby18}
\alias{KYDerby18}
\title{KYDerby18 Dataset}
\format{
A data frame with 144 rows and 8 variables:
\describe{
\item{Year}{Year of race. (Quantitative, discrete)}
\item{winner}{Name of horse that won. (Name, non categorical ID)}
\item{surface}{Surface of race course. ("dirt")}
\item{condition}{Condition of course on the day of the race, decided by race superintendent based on moisture content and physical state of racing surface. ("fast" or "muddy" or "slow" or "good" or "sloppy" or "dusty" or "wetfast")}
\item{starters}{Number of horses that actually raced (Quantitative, discrete)}
\item{PolePosition}{The position of the winning horse (ID of position)}
\item{Time}{Winning horse time of completion. (Quantitative in seconds.)}
\item{speed}{Average speed of winning horse during the race, calculated by distance (1.25 miles / 1.5 miles) divided by the time taken to complete the race. (Quantitative in MPH)}
}
}
\source{
Data compiled from official Kentucky Derby race charts
(1875–2018) published by Churchill Downs Incorporated and Equibase.
Original charts available at: https://www.kentuckyderby.com/history
and https://www.equibase.com/
}
\usage{
KYDerby18
}
\description{
Collection of all the winners of the Kentucky Derby from 1875 to 2018. Includes conditions, pole position, starters and speed. Important to note that the race distance changed in 1896 to the current 1.25 miles, where it was previously 1.5.
}
\keyword{datasets}
