% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeltingTimes.R
\docType{data}
\name{MeltingTimes}
\alias{MeltingTimes}
\title{MeltingTimes Dataset}
\format{
A data frame with 62 rows and 3 variables:
\describe{
\item{student}{Experimental unit identifier for paried data. (31 unique student identifiers such as TS)}
\item{time}{Time in seconds until chip completely melted. (Quantitative in seconds rounded to second.)}
\item{chiptype}{Type of chip melting. ("chocolate" or "butterscotch")}
}
}
\source{
Cal Poly Statistics Department. (n.d.). Lab 5: Exploration 2.1 – Chip Melting (Chocolate vs Butterscotch Chips) (Teaching dataset). California Polytechnic State University, San Luis Obispo.
}
\usage{
MeltingTimes
}
\description{
A teaching dataset from Cal Poly San Luis Obispo used to investigate melting times of chocolate chips and butterscotch chips.
Participants placed a single chip on the roof of their mouth and recorded the time (in seconds) to melt completely.
}
\keyword{datasets}
