% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piggrowth.R
\docType{data}
\name{piggrowth}
\alias{piggrowth}
\title{Pig Growth Dataset}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{Antibiotic}{Bernouli variable indicating if antibiotic is added to diet per day. ("Yes" or "No")}
\item{Vitamin}{Bernouli variable indicating if vitamin is added to diet per day. ("Yes" or "No")}
\item{Growth}{Average daily weight gain in pounds. (Quantitative, rounded to hundreth)}
\item{Treatmet}{Treatment group that each pig is assigned to ("Yes_Yes" or "Yes_No" or or "No_Yes" or "No_No")}
}
}
\source{
Iowa Agricultural Experiment Station (1952). Animal Husbandry Swine Nutrition Experiment No. 577.
}
\usage{
piggrowth
}
\description{
A scientist in Iowa was interested in additives to standard pig chow that might increase the rate at which the pigs gained weight. Two factors of interest were vitamin B12 and antibiotics. To perform the experiment, the scientist randomly assigned 12 pigs, three to each of the diet combinations (Antibiotic only, B12 only, both, and neither).
}
\keyword{datasets}
