% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLO2017.R
\docType{data}
\name{SLO2017}
\alias{SLO2017}
\title{SLO2017 Dataset}
\format{
A data frame with 454 rows and 16 variables:
\describe{
\item{price}{Price of home in hundreds of thousands. (Quantitative)}
\item{Bedrooms}{Number of bedrooms in house. (Quantitative, discrete.)}
\item{Sqft}{Square Feet of house, (Quantitative)}
\item{Sqft.1000}{Square Feet of house in thousands of feet.  (Quantitative)}
\item{YrBuilt}{Initial build year of the house. (Quantitative, discrete)}
\item{SqftxAge}{Interaction term of Sqare Feet x Age of house. (Quantitative)}
\item{std.sqft}{Standardized square footage of house. (Quantitative)}
\item{std.age}{Standardized age of house. (Quantitative)}
\item{std.sqftxstd.age}{Interaction of the standardized sqare footage and standardized age terms. (Quantitative)}
\item{garage}{Type of garage included in the house. Attached - A or Detatched - D. ("A" or "D")}
\item{Bathrooms}{Number of bathrooms in house. (Quantitative, discrete)}
\item{View}{If the house has a view or not. ("Y" or "N")}
\item{Pool}{If the house has a pool or not. ("Y"or"N")}
\item{SubType}{Type of residence / unit. Single Family Residence - SFR, Townhouse - TWNHS, Condiminium - Condo. ("Condo" or "SFR" or "TWNHS" )}
\item{LSqft}{Price of home in hundreds of thousands. (Quantitative)}
\item{Ac}{Acreage included with the property. (Quantitative)}
}
}
\source{
Tintle, N., Chance, B., McGaughey, K., Cobb, G., Rossman, A., Roy, R., Swanson, T., VanderStoep, J, Intermediate Statistical Investigations (ISI).
}
\usage{
SLO2017
}
\description{
A local realtor shared data with us on all list-ings in San Luis Obispo, CA, in 2017. We did some data cleaning (e.g., a home listed as 4200 square feet built in 1910 was excluded since online follow-up on that home could not confirm that interesting combination of values). A few other homes with some missing data were also removed. The data in SLO2017 contains list prices (in $100,000s) for 454 homes. Some quan-titative variables available for predicting the list price include square footage (in thousands of feet squared), number of bedrooms, and year built.
}
\keyword{datasets}
