% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Squirrels.R
\docType{data}
\name{Squirrels}
\alias{Squirrels}
\title{Squirrels Dataset}
\format{
A data frame with 18 rows and 2 variables:
\describe{
\item{Location}{Place where squirrel was found. ("Hemet" or "Big Bear" or "Susanville" or "Loophill")}
\item{Length}{Length of body in mm. (Quantitative rounded to mm.)}
}
}
\source{
Gerdes, N. E. (2001). Morphological and life history variation in three populations of golden-mantled ground squirrels (Spermophilus lateralis) in California (Master’s thesis, California Polytechnic State University, San Luis Obispo).
}
\usage{
Squirrels
}
\description{
A dataset with golden mantled squirrel body lengths and location found in California.
}
\keyword{datasets}
