% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stopping.R
\docType{data}
\name{Stopping}
\alias{Stopping}
\title{Stopping Dataset}
\format{
A data frame with 64 rows and 2 variables:
\describe{
\item{speed}{Speed of car in miles per hour. (Quantitative 4-45)}
\item{stoppingdistance}{Distance required for car to come to a complete stop. (Quantitative, in feet)}
}
}
\source{
Ezekiel, M. (1930). \emph{Methods of Correlation Analysis}. Wiley.
Original data recorded in 1920s; used in R dataset \code{cars}.
}
\usage{
Stopping
}
\description{
Data from 1920s about car stopping distances while traveling at indicated speeds.
}
\keyword{datasets}
