% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strawberries.R
\docType{data}
\name{Strawberries}
\alias{Strawberries}
\title{Strawberries Dataset}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{Firmness}{Force in newtons needed to pierce strawberry exterior. (Quantitative in newtons to tenth newton.)}
\item{Storage}{Enviroment strawberries were stored for two days before two days in room temperature. ("Air" or "ModifiedAir")}
\item{Variety}{Brand of berry tested. ("Allstar" or "Bounty" or "Kent" or "Selva" or "Vesper")}
}
}
\source{
Smith & Skog, “Postharvest Carbon Dioxide Treatment Enhances Firmness of Several Cultivars of Strawberry,” American Society for Horticulture Science, 27(5), 1992
}
\usage{
Strawberries
}
\description{
Dataset containing experiment data on strawberry storage.
Strawberries placed in different clamshells after harvest are left for two days in different environments, then in the same room temperature environment for two days. The "firmness" of berry was then tested using a probe to pierce the exterior.
}
\keyword{datasets}
